+++ package.json
@@ -18,9 +18,11 @@
     "test": "npm run build:packages && turbo run test",
     "lint": "turbo run lint",
     "wp-env": "npm run wp-env:start -w=wp/headless-wp",
-    "nextjs:dev": "turbo run dev --parallel --filter=!./projects/wp-multisite-nextjs --filter=!./projects/wp-multisite-i18n-nextjs --filter=!./projects/wp-nextjs-ts",
+    "nextjs:dev": "turbo run dev --parallel --filter=!./projects/wp-multisite-nextjs --filter=!./projects/wp-multisite-i18n-nextjs",
+    "nextjs:dev-app": "turbo run dev --parallel --filter=!./projects/wp-multisite-nextjs --filter=!./projects/wp-multisite-i18n-nextjs --filter=!./projects/wp-nextjs",
     "nextjs:dev-multisite": "turbo run dev --parallel --filter=!./projects/wp-nextjs",
     "dev": "npm run build:packages && npm run wp-env && npm run nextjs:dev",
+    "dev:app": "npm run build:packages && npm run wp-env && npm run nextjs:dev-app",
     "dev:multisite": "npm run build:packages && npm run wp-env && npm run nextjs:dev-multisite",
     "prepare": "husky install",
     "typedoc": "typedoc",
