+++ packages/core/src/react/components/index.ts
@@ -1,4 +1,5 @@
 export * from './BlocksRenderer';
+export * from './BaseBlocksRenderer';
 export * from './Menu';
 export * from './SafeHtml';
 export * from './HtmlDecoder';
