+++ packages/core/src/react/provider/DataFetchingProvider.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { FC } from 'react';
 import { SWRConfig, SWRConfiguration } from 'swr';
