+++ packages/core/src/react/provider/useSettings.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useContext } from 'react';
 import { SettingsContext } from './Provider';
