+++ packages/core/src/react/provider/useThemeSetting.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useThemeSettings } from './useThemeSettings';
 
 const get = (obj, path, defaultValue: any = undefined) => {
