+++ packages/core/src/react/provider/useThemeSettings.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useContext } from 'react';
 import { ThemeSettingsContext } from './ThemeSettingsProvider';
