+++ packages/core/src/rsc/components/BlocksRenderer.tsx
@@ -1,0 +1,12 @@
+import { BlockRendererProps, BaseBlocksRenderer } from '../../react/components/BaseBlocksRenderer';
+
+/**
+ * Server-only version of BlocksRenderer
+ *
+ * @param props Component properties
+ *
+ * @category React Server Components
+ */
+export async function BlocksRenderer({ children, ...props }: BlockRendererProps) {
+	return <BaseBlocksRenderer {...props}>{children}</BaseBlocksRenderer>;
+}
