+++ projects/wp-nextjs-app/src/app/layout.tsx
@@ -1,0 +1,24 @@
+import type { Metadata } from 'next';
+import { Inter } from 'next/font/google';
+import './globals.css';
+
+const inter = Inter({ subsets: ['latin'] });
+
+export const metadata: Metadata = {
+	title: 'Create Next App',
+	description: 'Generated by create next app',
+};
+
+const RootLayout = ({
+	children,
+}: Readonly<{
+	children: React.ReactNode;
+}>) => {
+	return (
+		<html lang="en">
+			<body className={inter.className}>{children}</body>
+		</html>
+	);
+};
+
+export default RootLayout;
